gui: refactor app::view::home module
